Move disable logic to lib
