+++ lib/autoupdate/agent.go
@@ -1,0 +1,132 @@
+/*
+ * Teleport
+ * Copyright (C) 2024  Gravitational, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package autoupdate
+
+import (
+	"context"
+	"errors"
+	"io/fs"
+	"os"
+
+	"github.com/google/renameio/v2"
+	"github.com/gravitational/trace"
+	"gopkg.in/yaml.v3"
+
+	"github.com/gravitational/teleport"
+	logutils "github.com/gravitational/teleport/lib/utils/log"
+)
+
+var plog = logutils.NewPackageLogger(teleport.ComponentKey, teleport.ComponentUpdater)
+
+const (
+	updateConfigVersion = "v1"
+	updateConfigKind    = "update_config"
+)
+
+// UpdateConfig describes the update.yaml file schema.
+type UpdateConfig struct {
+	// Version of the configuration file
+	Version string `yaml:"version"`
+	// Kind of configuration file (always "update_config")
+	Kind string `yaml:"kind"`
+	// Spec contains user-specified configuration.
+	Spec UpdateSpec `yaml:"spec"`
+	// Status contains state configuration.
+	Status UpdateStatus `yaml:"status"`
+}
+
+// UpdateSpec describes the spec field in update.yaml.
+type UpdateSpec struct {
+	// Proxy address
+	Proxy string `yaml:"proxy"`
+	// Group update identifier
+	Group string `yaml:"group"`
+	// URLTemplate for the Teleport tgz download URL.
+	URLTemplate string `yaml:"url_template"`
+	// Enabled controls whether auto-updates are enabled.
+	Enabled bool `yaml:"enabled"`
+}
+
+// UpdateStatus describes the status field in update.yaml.
+type UpdateStatus struct {
+	// ActiveVersion is the currently active Teleport version.
+	ActiveVersion string `yaml:"active_version"`
+}
+
+// Disable disables agent updates.
+// updatePath must be a path to the update.yaml file.
+func Disable(ctx context.Context, updatePath string) error {
+	cfg, err := readUpdatesConfig(updatePath)
+	if err != nil {
+		return trace.Errorf("failed to read updates.yaml: %w", err)
+	}
+	if !cfg.Spec.Enabled {
+		plog.InfoContext(ctx, "Automatic updates already disabled")
+		return nil
+	}
+	cfg.Spec.Enabled = false
+	if err := writeUpdatesConfig(updatePath, cfg); err != nil {
+		return trace.Errorf("failed to write updates.yaml: %w", err)
+	}
+	return nil
+}
+
+// readUpdatesConfig reads update.yaml
+func readUpdatesConfig(path string) (*UpdateConfig, error) {
+	f, err := os.Open(path)
+	if errors.Is(err, fs.ErrNotExist) {
+		return &UpdateConfig{
+			Version: updateConfigVersion,
+			Kind:    updateConfigKind,
+		}, nil
+	}
+	if err != nil {
+		return nil, trace.Errorf("failed to open: %w", err)
+	}
+	defer f.Close()
+	var cfg UpdateConfig
+	if err := yaml.NewDecoder(f).Decode(&cfg); err != nil {
+		return nil, trace.Errorf("failed to parse: %w", err)
+	}
+	if k := cfg.Kind; k != updateConfigKind {
+		return nil, trace.Errorf("invalid kind %q", k)
+	}
+	if v := cfg.Version; v != updateConfigVersion {
+		return nil, trace.Errorf("invalid version %q", v)
+	}
+	return &cfg, nil
+}
+
+// writeUpdatesConfig writes update.yaml atomically, ensuring the file cannot be corrupted.
+func writeUpdatesConfig(filename string, cfg *UpdateConfig) error {
+	opts := []renameio.Option{
+		renameio.WithPermissions(0755),
+		renameio.WithExistingPermissions(),
+	}
+	t, err := renameio.NewPendingFile(filename, opts...)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	defer t.Cleanup()
+	err = yaml.NewEncoder(t).Encode(cfg)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	return trace.Wrap(t.CloseAtomicallyReplace())
+}
