+++ lib/autoupdate/testdata/TestDisable/already_disabled.golden
