+++ lib/autoupdate/testdata/TestDisable/enabled.golden
